Fix the file structure
Constants and config objects are separated into their own modules
